Don't disrupt SUBFRAME on CONTINUE hack
For a construct like REDUCE-EACH it can be in the middle of
walking through an evaluation, but wants to run some code like
a loop body.  So it uses a trick to temporarily set the executor
of the subframe to be a "Just_Use_Out" executor, that will pass
the body result back through a frame pushed above it.

It's a neat trick, but... the system automatically unwinds the
frame stack on throws.  So if the body block does a CONTINUE
that was destroying the state of the enumerating frame.

Not clear how to fix this in general, so for now if a frame
with the Just_Use_Out executor does a throw, then it will keep
itself on the stack.