Lose DEBUG_ENSURE_FRAME_EVALUATES
The concept behind this debug flag was to catch cases where the
evaluator wasn't being given control, e.g. the bypassing of empty
blocks in a loop or parse.  This would lead to not being able
to use Ctrl-C to cancel because the polling for the cancel signal
was in the evaluator loop.

So the flag was initialized to zero, and then set if the evaluator
ever ran.  But it was disabled due to problems, and it also does
not have as much of a point now that all evaluations are expected
to go through the trampoline.