Don't pass VOID to predicates in COMPOSE
It *may* be desirable to do so, but only in cases that are able
to handle voids.  It seems wrong to make it so that the dispatch
of functions that don't take voids should unilaterally return
void instead of error when asked to run, so it's a tricky issue.

But certainly the general case of a "normal" parameter should just
not be offered the void.