Change evaluator bias to run to end by default
Initially there was an executor-specific flag for the evaluator
called EVAL_FLAG_TO_END which told it to keep running until the
end was hit.  But that is the much more common intent, to where
it makes more sense to make EVAL_FLAG_SINGLE_STEP the special
flag.  Most of the cases running single steps explicitly push
a frame already--so they have a place to set that flag.