Fix test for Is_None()
A flawed test was looking to see if something was the none isotope
by seeing if its ID was SYM_0

But SYM_0 doesn't mean a null label, it means a symbol ID that is
not in the default set.  So this would treat anything like ~asdf~
as if it were a none, when only ~ should be.