Delegation in non-ACTION! frames, group branches
This creates a separate executor to handle branches that are
GROUP!s.  That's needed to solve the problem of making stackful
group evaluations as a first step in branch evaluation.

It also introduces the idea that delegation is something that
can be used in executors as well as in action dispatchers.  This
makes a new R_DELEGATION return result.  If this is returned to
an action dispatcher, it will not Drop_Action() at that time, but
when the subsequent frames finish...because mechanically actions
should stay running for conceptually stacklike needs.