Address action executor Recycle/Torture issues
The Chainer_Dispatcher chose to preserve a frame and reuse it across
several distinct actions.  This meant the RECYCLE/TORTURE code could
hit a recycle across a trampoline bounce where an action frame had
been dropped--so it still had the Action_Executor but not the valid
properties for an executor.

This could be handled by saying that the frame has a state byte that
is not ready for being seen by the GC but still has Action_Executor,
or by making the executor something else.  This makes it nullptr in
the Drop_Action() code, so if the frame is seen by GC it will not
try to mark the action information.