Implement old Do_Branch_Throws() via continue frames
The continuation code has matured to where places that used to call
plain Do_Branch_Throws() can just push a continuation frame and
then call the trampoline.

It's not as good as writing code to be stackless, but it's better
than having two copies of the code for branches that does the same
thing.