Take advantage of continuation arg/output equal
The idea that a continuation for a function branch can take an
argument that's the same as the output location is useful, in
particular when a predicate wants to put its output in the same
location that its input would be when not processed by a
predicate.

This makes code that is shorter, clearer, and more optimal.