Beginning of Stackless in JavaScript extension
This turns off the Asyncify feature in JavaScript in favor of use of
the stackless methodology now embraced by the system.

It will not work in cases where native code is not stackless at the
moment a JS-AWAITER is invoked.  But those cases are being looked
into aggressively, and it's best to just go ahead and embrace this
since it's the way to go.

The size of the wasm.js goes from 1.8mb down to 1.1mb as a result.

Also drops the debugger extension from the main build as it is
completely broken at the moment and must be completely rewritten
for stackless.