Attempt simplification of port dispatch code
The extremely half-baked ideas behind port dispatch from R3-Alpha
are a challenge to move to stackless (standing on the shoulders of
the also extremely half-baked generic dispatch concept).

This attempts to simplify a few of the more unnecessary steps in
the chain, some of which were added by Ren-C with good intent but
that offered no closer solution to the problem.

Common handled actions by all contexts are done by falling through
to calling REBTYPE(Context) instead of a strange other function.

The Do_Port_Action() code is put in one place, shared by virtue of
using R_CONTINUATION to bounce a generic dispatcher that has
changed its first argument to become a PORT!