Make REFLECT a GENERIC again
The needs of being able to say TYPE OF X where X is NULL and get
back NULL motivated an entry point to uses of REFLECT that were
not based on type dispatch.

(This also allows for generic reflection, because otherwise every
type would have to respond to TYPE OF which is repetitive.)

With continuations it's a bit easier to reuse that logic and
actually declare REFLECT as a generic.  There is a native form
that is kept, which could be used to overwrite the generic and
add the features of null tolerance and generic reflection...but
for the moment those abilities are kept in OF.  REFLECT-NATIVE
remains in case the idea of overwriting REFLECT in LIB seems
to be a good idea later.