Rethink Redo_Action
This brings an incremental improvement to the Redo_Action mechanic
used by HIJACK and the port machinery, which is able to line up
parameters for functions that were not derived from each other...
by position for non-refinements, and by name for refinements.

It still uses the technique of turning ordered normal parameters
and refinements with arguments into an array.  The array of inputs
are then QUOTED! to prevent double-evaluation.  These techniques
only work on simple cases, but still it's far better than R3-Alpha.

More work should be applied as time permits.