Mark port dispatch very slightly stackless
This makes the dispatch to usermode functions of things in port
actors like READ stackless.

(Important e.g. for the JavaScript extension, because that READ
actor function is exactly where it wants to run a fetch(), which
requires being able to suspend the Rebol stack in Wasm to give
control back to the browser to service the request...and then
resume the stack once the request is ready.)