Push Continuation API
This offers rebPushContinuation(), which isn't really an API that
is good for users but helps bridge the desire to write variadic
function calls in a style that actually gets their own frame to be
used in making stackless natives.

Further work will be needed to come up with a model for plain C
programmers to use such APIs, but this is good enough to make DO
of a FILE! or URL! stackless.  So it's applied to that.