Stackless implementation of the CONSOLE function
The console was designed not to do I/O from within its logic, but
to submit everything through EMIT requests that might potentially
cause an error.  This allows it to sandbox problems that affect the
UI and revert to a default console state.

However, those I/O requests need to be stackless if CONSOLE is to
be used stacklessly (in something like the web build).  This uses
the rebPushContinuation() API to get there.

As a result the CONSOLE implementation is less independent of the
%sys-core.h internal API than it had been.  Hopefully this can be
revisited and a model for C programmers to program asynchronously
in a way more pleasing than being subject to the internal API can
be created.