Write resolve/reject directly to frame's OUT
This removes the SignalResolve/SignalReject indirection that
was needed from Emscripten.  That put an entry into a table
held by frame_id and then merely notified the JS-AWAITER
that a value could be preoduced (from a JavaScript function
that had been stored).  The table was then queried later.

Instead, this uses the fact that frame_id can actually be
made into a frame pointer to have the Resolve/Reject helpers
write the data directly into the JS-AWAITER frame's output
cell.  It doesn't need to pollute the table with entries,
and the JavaScript_Dispatcher() can proceed use the value
without extra steps when it resumes.