Cohere EVAL_FLAG_MAYBE_STALE handling
The Evaluator_Executor() had code for removing the stale bit,
but that was already being done at the trampoline level.

Also moves asserts on STATE = 0 that if your input is not
void, you need to know you're dealing with potentially
stale input and need the MAYBE_STALE flag.

Doing this required several changes, and in the process it
gets rid of the non-stackless Do_Feed_To_End_Throws() and its
awkward handling of issues prior to the EVAL_FLAG_SINGLE_STEP
distinction.