Tighten up abrupt failure logic
This rethinks abrupt failures such that every abrupt failure
can work as a "definitional return" from the function that
raised it.

The logic is clarified a bit, and it's enforced that once a
frame has encountered an abrupt failure that it handled, it
cannot continue and must rethrow the error.