Major New Feature: Definitional Failures
This feature changes it so that FAIL does not fail "in the
moment" but generates a failure state (what might be thought
of as an "error isotope") whose ^META state is an ERROR!.
The "in limbo" failure state will raise an alert if it is
not processed, but can be ^META'd in order to get processing.

This feature allows for a difference between FAIL and
RETURN FAIL, where arbitrary failures are not connected to
a particular function invocation...while the definitional ones
are.  This makes it possible to respond to errors that you
know are coming from a particular function invocation.