Improve safety of Meta_Unquotify()
It should be fairly difficult to create instances of the
Failure state cell--and they should generally pass through a
frame that has explicitly said that it's interested in getting
a failure back.  There are too many Meta_Unquotify() instances
that can make them, so this removes it from their default at
this lower level.

The UNMETA( ) specifically targets output cells and so it is
a safer thing.