Clarify Trampoline Entry Points
Given the common pattern of "push frame, trampoline, drop
frame" this embraces that for Trampoline_Throws(), which uses
a more general version that uses the current top frame but
treats it as a root.  This all stands on top of the most
generalized version that runs from whatever the top frame
is whether it's a root or not.

The names of the functions are extended to detail the
differences in semantics.