Eliminate EVAL_MASK_DEFAULT
A concept that was thought of was to force all those who
were initializing EVAL_FLAGs on a frame to include an
EVAL_MASK_DEFAULT.  This would make it possible to negate
the sense of certain options, so the "unused" state would
be a true bit.

This wound up creating a hassle of responsibility for
knowing if anyone in a flag compositions context had done
the inclusion of EVAL_MASK_DEFAULT or not.  It also made
all the sites for initializing flags markedly longer.

As of yet no truly interesting optimization has been made
using the positive sense of an unused EVAL_FLAG, so this
is really just making the source more cluttered.  Backing it
out and just having EVAL_MASK_NONE as an alias for 0 to use
when no flags are being applied (to know what the slot is
representing--flags) is much better.