Allow pushing actions into unpushed frames
There was an order of behavior that would first push a
frame to the frame stack, and only then allow the frame
to be filled with information describing the actions
fields and allocating memory for its arguments.

If anything, this is backwards--you'd rather do the frame
filling and then push it.  But there's not really a reason
to restrict it to one way or the other.

But the key motivation for changing it is that logic in
Push_Frame() is about to take over the erasing of the output
cell if the frame flags do not carry EVAL_FLAG_MAYBE_STALE.
There may be information in the cell being used as output.
that could be used as an input when building the frame.