Make Push_Frame() automatically RESET() OUT
There were a number of assertions in place trying to keep
callers on the right side of the rule that you must have a
void (or stale void) input when calling into executors for
their initial state...unless they are willing to tolerate
stale values.

(One reason for this was a misguided belief in trying to
optimize so that already-cleared cells wouldn't need to be
cleared again.  This is less relevant than ever, as a cell
reset to become void is a single masking operation on a
platform-pointer-sized integer of bits.)

This does some juggling to make it so that Push_Frame()
itself does the RESET() to a void state if the MAYBE_STALE
bit is not set in the frame flags.  It requires that some
cases that know the output value won't be stale use the
flag in order to suppress losing the output they know they
are going to keep...but this seems an acceptable price to pay.