Improve safety of executor-specific flags
To save on space, increase locality, and make it easy to flip
a lot of flags at once... frames pack in flags to a single
platform-integer-sized slot at the front of their struct.

Some of these flags apply to anything running through the
trampoline, while others are reserved for a specific executor.

This makes it so the executor-specific flags have a debug
check that the executor you think is in effect actually is.