Clarify RUNNING_ENFIX frame flag
Due to the scarcity of frame flags, a common RUNNING_ENFIX
flag was recycled for two distinct purposes in action frames
vs. evaluator frames.

That recycling is just confusing now, and the evaluator case
can be handled using data in the evaluator union state vs.
a scarce frame flag.

This clarifies the situation in order to advance the reclaim
of generic executor flags.