Merge "multi return" executor state to eval
The previous formulation of the `u` union inside of the
Reb_Frame did not make it clear that these were per-executor
and not per-random-operation.

Hence using the space to store a dsp_circled could overwrite
the cache of the currently processed item.

The eval executor actually has a lot of spare space, so
it can put the dsp_circled in the state (if it were to be
a union it would need to be a subunion)