Changes so that NEWLINE_PENDING is executor flag
This pushes through what it took to migrate newline_pending
from being a boolean in the executor state union of the
frame, to being an executor flag in the header.

It turned out to be fairly involved, because the frame must
be kept alive on the frame stack.  This goes ahead and makes
that change, and also removes the /RELAX option until it
can be rewritten in a way that makes sense for stackless.