Get rid of Scan_To_Stack()
This flattens out the calls to the Scan_To_Stack() function
so that commonalities between the stackless and stackful
callsites can be factored out in a better way.

A particular helper routine in Fetch_Next_In_Feed() needed
to be moved into a .c file due to the new dependency of the
scanner on things like DECLARE_END_FRAME().  The reason the
code was in a header was for inlining, and that portion would
not be good for inlining anyway