Make many of scanner's errors definitional
This is a beginning step in making errors that would be
raised during scanning "definitional" such that they can
be caught by ATTEMPT or EXCEPT, vs. just the TRAP (that
is origin-agnostic about errors, making it nearly useless
for program logic purposes).

Going further would require making Locate_Token actually
able to return a definitional error, and doing an audit of
the service routines that are called by the scanner...which
is necessary but will involve a lot more work.