Make DIDNT_QUOTE_LEFT_PATH an executor flag
It's not obvious that both EVAL and ACTION versions of the
flag are necessary, but also not obvious that it isn't...
but clearly it's not relative to composer executors or the
scanner executor, etc.

So this moves them into praallel EVAL and ACTION executor
flags, freeing up the space for other executors and reclaiming
a common flag.