Improve readability of data stack operations.
The data stack operations have historically been prefixed as
DS_TOP, DS_PUSH(), DS_DROP(), DS_AT() and DS_DROP_TO().

This change goes ahead and bites the bullet on TOP, PUSH()
and DROP() letting them be defined just as words with no
prefix, where the data stack is implied.  But for DS_AT and
DS_DROP_TO it makes the call that these probably are fine
to write out long form, as Data_Stack_At() and
Drop_Data_Stack_To().

Reason being that something like Drop_Data_Stack_To() usually
sits on its own line anyway, so brevity and not interfering
with the context with the DS_ "noise" isn't as important.

In the case of AT(), that simply doesn't stand alone the way
that PUSH() and TOP can.  So it needs some context, and the
abbreviation isn't very appealing.  It's not used all that
often so calling it Data_Stack_At() isn't so bad.

This also changes the name of STKVAL(*) to StackValue(*)