Re-style Symbol as a Type Macro
By defining something like Symbol pointers via a type macro,
e.g. Symbol(*) or Symbol(const*), greater flexibility is
afforded in the C++ build to be able to make these actual
objects that have some kind of compile-time checking.

After weighing the issues deeply, the "oddness" of doing
this vs. plain `Symbol *` or `const Symbol *` is outweighed
by the advantages, so this is going to be the strategy taken
with the newly named types, as the REBXXX heritage is
dropped.