Stylize Code to Support C++ Exception Handling
This rethinks the handling of errors in the trampoline so that
the code for handling the errors is bracketed in blocks, in the
style of a `try {...} catch(e) {...}` type of construct.

By doing so, it offers the ability to build the codebase using
structured exception handling.  While this involves more baggage
on some platforms (e.g. those you would typically use in embedded
C programming), other platforms support it natively--where it may
be difficult or impossible to emulate the behavior of arbitrary
C setjmp()/longjmp() code.