Make STATIC_ASSERT() usable in %reb-config.h
It's a pain to checks with #if + #error + #endif

Though it's an opportunity to provide a better error message,
anyone who gets these errors probably will need to go read
the place where the error happens for comments that give
context anyway.  Better to do it in one line vs 3.