Add interim defined? and undefined? functions
The question of if a variable is "set" or "unset" has become
something fairly vague in the Sea of Words model, and we
don't want to give misleading answers.

While the direction of Sea of Words is good, there are still
some places where the question of if a definition exists is
useful and actionable, regardless of variable status.  This
introduces a couple of helpers to bridge the gap while the
broader model is emerging.