Add "WASI" (WebAssembly System Interface) Target
This makes a new target that can compile WebAssembly to
run in containers like Wasmtime, Wasmedge, or Wasmer.

It uses the wasi-sdk toolchain.

While the console and command-line processing work, the use
of libuv means that the filesystem and network APIs are higher
level than the basic I/O provided by the C runtime library
used by wasi-sdk (a.k.a. wasi-clib).  This library also lacks
any network facilities--even basic ones--at this time.

(Rust libraries are further ahead in bridging to the new
WASI network features, and are leveraged in WasmEdge examples,
including a JavaScript interpreter that is hooked into the
Rust routines for file requests.)