Rename REB_R to Bounce
Executor results have been called REB_R for "Rebol Result".
It's desirable to get rid of the REB_ part, but Result is
too generic for a global type name.

Bounce is a little bit weird, in particular because not all
results actually seem to "bounce".  For instance, returning
a void through BOUNCE_VOID will drop the stack level and
consider the call complete.  However, that is "bouncing"
in a sense...it's just bouncing a level off the stack...so
bouncing lower.

It starts looking normal pretty quickly...especially when
compared with REB_R.