Move SYS to SYS.UTIL, SYS as SYSTEM synonym
Historically a set of helper functions for the system were
in a context called SYS.  This had things like sys.make-scheme

A different object--a "system object" was in SYSTEM, where
things like system.script.args were kept.

This was a confusing division.  So this change moves the
helper functions into SYS.UTIL, and makes SYS a synonym for
the system object.