Re-enable Ctrl-C in the console
When the console was converted to stackless, it had to use
continuations.  This meant it was not able to use the
rebXXXInterruptible() form of the API, because it was needed
that the console would yield control from its REBNATIVE()
function (otherwise it wouldn't be able to do things like
modify the DOM in the web build as part of the script it
was executing).

But the rebXXXInterruptible() functions were what fiddled
the mask of flags that told the Trampoline when it was okay
to respond to a Ctrl-C signal.  Without that, it assumed that
it was not okay.

This puts that fiddling directly into the CONSOLE native,
before the request for the continuation is run...and then
turns the handling for Ctrl-C off after returning.  This way
you cannot cancel the console implementation itself--only
the code that it runs as part of the skin, or a user eval.