Cleanup optional preallocation of series nodes
The idea of being able to use a global or fixed buffer to
store a series node came up in order to have quickly
accessible locations for the default declarations in Lib
(like Lib(NULL), etc).  But it's more generically useful,
so there's a kind of "emplacement" (C++ term) where the
construction is done directly into memory locations.

This does a couple of adjustments to that, to generalize
requesting GC when a series node allocation trips the
ballast level, and to make the types representative of
when the criteria for "REBSER" or "REBARR" have not been
met, so it still passes the memory buffer as a void*.