Warn instead of error regarding REBOL_EXPLICIT_END
While not technically in the standard before C++11 or C99,
variadic macros had somewhat wide support.

As implicit ends are required to build the core now, it is
reasonable to relax the error to a warning, in case someone
wants to try building the core as C++98.  (It works with
Emscripten, for instance.)