Don't define Reb_Node as struct in C++98
Ren-C has to bend around a bit in order to make a proper
subsetting class heirarchy out of the node and cell types.
But in the C build it just throws up its hands and says
nodes are void* as the base class of Reb_Series and Reb_Cell

C++98 does the same thing, to avoid having to tailor the
class heirarchy with exceptions for missing features that
came about in C++11.

But libRebol in the interface wasn't noticing this and
trying to make the base a Reb_Node struct, not a void*,
in all C++ builds..including C++98.  This adds in the
requirement for C++11 to make the struct.