Support nullptr when already shimmed
The Emscripten build makes a shim for nullptr in order to
compile C++98 codebases.  If we try to do it too, then it
will create an error.

This assumes anyone who has defined nullptr made a compatible
version, and makes the place where the instance is declared
more clear by using a #define.