Use m_cast_helper in all C++11 builds
const_cast<> cannot be used on smart pointer classes.  As
a result, some helper is needed to "unpack" the constness
from the wrapper.

Previously the wrapper was only present in DEBUG_CHECK_CASTS
builds, but this means the one specifically for mutable
casts is needed in all builds that would use smart pointer
classes.

One result of this is that the DEBUG_UTF8_EVERYWHERE mode can
now be used in non-DEBUG_CHECK_CASTS builds.