Add SeriesHolder with DEBUG_COUNT_LOCALS test
This creates a new mechanism to use with typemacros, called
the SeriesHolder.  This does reference counting on the
series that are marked with Series(*), so that when a
GC_Kill_Series() happens it can assert that the cont is
at zero (e.g. no local stack references are outstanding
which might consider that pointer live).

As a first test, this applies the methodology to string,
which is renamed from Reb_String to Raw_String to help point
out that a String(*) has this protection while Raw_String
does not.

Eventually this could grow into features that let the C++
build do things like run a GC at an arbitrary moment besides
just between evaluations (e.g. to reclaim memory at the
moment an out of memory occurs).