Relax restriction on C++ in internal headers
The external API has gained a lot of power, and is aiming to
be the main API used by extensions.

With String(*) becoming a C++ type, it joins the Utf8(*) types
in needing C++.  For now, the rule on internal headers being
extern "C" only is restricted to the C build.