Get rid of virtual binding index limit
An early concept in virtual binding was to make it so that
if new fields were added to an object after the fact, it
would try to remember how many fields were in at the time
of the virtual bind.

There are bugs in this exposed by the SeriesHolder.  Fixing
the bugs isn't worth it, given that there's an impending
redesign of binding and modules--now that some experience
with Sea of Words has been had.