Make Value a TypeMacro(*)
While the approach was being evaluated, some small steps were
put in the code to see if just plain "Value" was a good pick
for the name of value cells.

It seems to be, and further it seems a good idea to keep
pushing on the use of typemacros to be able to add C++ alternate
compilation with smart pointers.  (Whether these ever add
features or are just used for debug tracing remains to be seen,
but there are benefits which have already become apparent.)