Rename REBSIZ to just Size, minor cleanups
This changes the name of the size_t alias from REBSIZ to Size
and dabbles a little bit with a coming idea of distinguishing
Index (1-based) from Offset (0-based) in the type system.

A few places that were not using option() are redone to use
it, and also simplifies the length calculation of ANY-WORD!s
to use normal UTF-8 enumeration.