Fix GET-GROUP! parameters
Rather than call generic evaluation, the processing for
GET-GROUP! arguments went through the more direct getting
of variables.  However, GET lacked support for getting a
GROUP! when the legality of getting paths and tuples with
groups in them was made.

This quickly patches that so that GET-GROUP! branches work
(primarily because it's the topic of a forum post).