Restyle DECLARE_FRAME and DECLARE_FEED macros
At once point, frames and feeds were C-stack-allocated.  The
DECLARE_FRAME and DECLARE_FEED macros were something of an
attempt at streamlining that and making it "more optimzed".
They would declare the memory via a named variable on the
stack, and then they would declare a pointer which would
be assigned the address of that variable.

With stackless, frames and feeds are pool-allocated.  So
this really just obfuscates things...and more clarity is
given by typical variable declarations.

This changes them to that style.