Change Any_XXX terminology to XXX_Cell
There was a terminology attempt to distingush routines that
were supposed to operate on values as leveraging the ANY-XXX!
convention.

With the usages of the term "Cell" growing in the codebase,
it makes more sense to say things like:

    Init_Binary_Cell(out, Make_Binary(...))

This way binary can be reserved to mean the Binary(*) series
type, while "Binary_Cell" can refer to the implementation of
a value that has an index and binding etc.

Here that is started with the Any_XXX terms.