Alloc/Free_Node() => Alloc/Free_Pooled()
Terminology has moved along, though somewhat inconsistently,
toward the idea that the family of pointers in which both
"Series" and "Cells" fit be called "Nodes"

This leaves a little bit of a question about what to call the
little size-of-2-cells stub that is allocated as a tracking
entity for series.  This has been called "Series Nodes" for
some time, and sometimes thus being confused in the dichotomy
between "Cell" and "Node"...when the goal is for Cell to be
a distinguishable pointer in the same base class.

In any case, Alloc_Node() and Free_Node() do not do that...and
were just pooled allocations, sometimes returning things like
blocks of memory for series data.  They return void pointers,
so it's not like they return any particular class.

Alloc_Pooled() and Free_Pooled() seem like good enough names.