Change return_value() to return COPY()
Following the spirit of not hiding the `return` under an
abstraction, this does so even with the case of a wrapper
for return that includes "return" in the word.

The result is a macro that has the raw keyword return in it
so it's more obvious, and that still has added checks and
the implicit frame_->out parameter.