Change return_branched() to return BRANCHED()
This is the last of the `hiding return behind a macro`
instances that is done a better way.

BRANCHED() is just an assert--and it's tempting to explain
that at every callsite, but better to have people jump to
the declaration and read that's what it does vs. add more
noise to the code.