Usermode SCRIPT?, delete FIND-SCRIPT native
There was a fair bit of C code written to the purpose of
trying to recognize `Rebol [...]` at the header of a script.

Routines Scan_Head() and Scan_Header() were there to support
the FIND-SCRIPT native, which had custom entry points to
the scanning code--making it difficult to change the
interface there.

This moves that logic into usermode, where it can be more
easily designed and considered--with much less code.