Avoid scanner recursions for Fetch_Next_In_Feed
Little had changed since the initial grafting of making the
scanner able to splice values together with scanned text
in the API.  And it needs changing.

The basic idea was that scanner overhead was avoided by
only starting up a scanner once a UTF-8 character string
pointer was detected in the variadic input.  At that point
it would load the remaining input as an array.

But the work in processing the variadic items was too
complex to repeat in the scanner.  So it turned around and
called *back* into the detection itself--creating a very
confusing situation.

This takes a basic step of just factoring out the handling
for variadic items such that it's not so bad to have two
places that do it--the basic Fetch_Next_In_Feed() of the
variadic loop, and the internal logic of the scanner.

It also coalesces all the entry points to the scanner so that
they go through the variadic interface, instead of having
a separate model where there is no feed.  (This goes along
with the fact that every scanner level has a stack frame,
so they have to have a feed anyway.)