Begin abstracting feed->value to feed->p
The longstanding concept of having feed->value be preloaded
without needing to be first Fetch'd is being challenged by
the idea that variadic feeds need to defer any erroring
until a guard is actually pushed by running the trampoline
of the frame the feed is being constructed for.

This means the notion of "current pointer" will have to be
initialized as a const void*, and hence it will not
necessarily be a valid Cell(*) prior to the first fetch.

In order to assist in this abstraction, a short name that
will be used for casting the pointer to a cell is picked
as `At_Feed()`

Feed_Value() is longer and also sounds like it takes some
action (like "Advance value, feed it forward").  The idea
of `Feed_At()` looks like it might return a Feed, like "give
me the feed at this location".

The odd-looking At_Feed() looks pretty good for being
something along the lines of "get me what's at the feed in
the current moment", a sort of dereferencing operator.

(This doesn't actually change any of the program logic, it
just prepares by making feed->value a const void* called "p")