Finalize variadic feeds if replaced by UTF-8 scan
UTF-8 scans were replacing the variadic feed with a block,
but not finalizing them (e.g. not calling va_end()).  It's a
no-op on most platforms, but not all...so calling va_end() is
a requirement.

Also make the GC reification code aware that the variadic
status can change across Fetch_Next_In_Feed().