Reuse Frame(*) in SET-BLOCK! used in eval step
Some code was established to be used by "reframers" which
allowed the capture of a FRAME! representing a function
invocation in the evaluator--which didn't actually run
the function.  This was reused in the first implementation
of a SET-BLOCK! multi-return.

However, that would tear down the Frame(*) that was used and
extract the variables, requiring a new frame to be pushed
and configured to execute if it was going to be invoked
later.

This factors out action execution to give back the frame
before throwing it away, as a first step in crossing over
the work done by SET-BLOCK! and generalized return features
supported by the evaluator.