Mark pushed refinements using common routine
Refinements are going to be pushed on the stack mixed with
variables that need to be written back at the end of an
action call, so these will have to be different.

It's best to encode that difference via a common name vs.
spreading the decision of what datatype to use at each
callsite.