Action dispatching status as frame flag
Historically frame flags were scarce.  (They still are, but
are slightly less so with executor-specific flag separation.)

This meant that if it were possible to encode a state without
using a flag, it was done.  So the action enumeration state
in frames was examined to see if it was at the "end", to
see if it was "fulfilling" or not.

But this was wasteful and slow... it required fetching two
enumeration variables (f->key and f->key_tail) and comparing
them, when a flag would be more convenient.  This also hit
a snag with functions with no arguments and no returns (e.g.
a lambda), because there's no way to tell if they have
already been through a dispatch or not (important in the
stackless model).

This goes ahead and reclaims a flag for the purpose, from
something low-priority (but leaves the low priority flag
around as documentation of the issue, just sets it to 0
so setting it is a no-op and it can never be found set).

With this, it's also possible to reclaim much of the state
of an action frame's union variable during dispatch.  The
idea needs more development to do reuse that space cleanly,
but this makes a start by "trashing" the key, key_tail, arg,
and param fields when not needed.