Auto-dropping thrown frames only in Action_Executor
In the first draft of stackless, returning thrown was
considered a "cooperative abort", and it was required that
every dispatcher (e.g. natives) drop any pushed frames
before they returned.  This made it annoying for things
like ANY or CASE that had throws happen during evals.

So the trampoline dropped subframes for them.  A problem
was noticed with this in that those stacks may capture a
baseline higher than what the action executor would be
dropping on its own exit.

Because the users of this questionable service are all
dispatchers called by the Action_Executor() at this time,
the feature of dropping frames on throws is moved to be
a feature of that executor and not the trampoline as a
whole.  It would need to be thought out more if it were
to be used by something like the Evaluator_Executor() or
the Composer_Executor(), for example.