Some reorganization of %c-action into sections
The new concept is to do commenting more in blocks with
numbered references, so that reading the code isn't interrupted
as much.

The stackless style also means dividing into scoped sections
is more frequently possible since local variables aren't
supposed to carry over as much.  This has yielded a fairly
nice style of:

} label: { ////////////////////////////////////////////////

Where the begin and end of scopes for each section looks natural
on the same line as the label.