Inefficient first-cut at multi-return proxying
This starts to change the semantics of multi-returns to where
a multi-return variable can be read and written normally
(e.g. not with GET and SET).  The actual getting and setting
is done by an epilogue run by the function handlers.

This just inefficiently walks the frame at the end of each
function call, in order to get the semantics right.  Better
approaches would blend this in elsewhere (e.g. the type
checking pass, using stack material pushed from enfix).