Eliminate STACK_NOTE_METARETURN
Implementing SET-BLOCK! behavior for multi-return came with
a desire to not concern the called function with the multireturn
properties (like desire to circle or ^META), and just pass the
variable to set.  Hence the variable was planified and the
information was not carried by it...forcing the information for
whether it needed to be ^META'd to be a flag on stack variables.

The new concept is to fully embrace proxying, where the user
of the multireturn function does not even know the name of
the variable...much less if it is being ^META'd or is the
circled output.  This uses a meta variable as the reified
form of the intent.