Fix TUPLE! action-tested-first optimization
VAL_ACTION() now errors if asked on isotopes.  However, the
TUPLE! handler in the evaluator deliberately asks if the
result is an action before testing if it's an isotope to try
and squeeze a tiny optimization out.

This changes it to use VAL_TYPE_UNCHECKED() == REB_ACTION to
preserve the optimization and dodge the assert.