libRebol-based extensions (no %sys-core.h)
A conflict between MacOS CoreFoundation and %sys-core.h in
the definition of "Size" created the first unfixable
collision between %sys-core.h and a platform API.

To address the issue, this makes a small move in the
direction of being able to write extensions using only the
public API, and applies it to the UUID module.  This makes
the MacOS build complete again.