Make Bounce always REBVAL (for now)
The attempt to create a native extension that didn't know
about the internal Bounce type revealed a problem.  Even
though Bounce is a standard layout struct containing a
single REBVAL*, it was not acting compatibly with functions
returning a non-struct-wrapped REBVAL*.

Working this out is not a particularly pressing issue, so the
Bounce type can just be for documentation.  Review when there
is time for this particular detail.